requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
requests
werkzeug.security
